Add CompositeSetup#more_like_this_fields
Sunspot#more_like_this searches were failing if multiple types were provided
to run the search over. This is because a Sunspot::CompositeSetup object is
used to extract the fields in this case, instead of the usual
Sunspot::Setup object. The CompositeSetup class was missing the required
more_like_this_fields method, which is added by this commit.

A test is also added to the more_like_this integration tests.

Related Story:
  http://www.pivotaltracker.com/story/show/45804697
  ("Recommended for you showing strange results.")

Change-Id: I5c839ff20b39e98098321312ef652de753091916
Reviewed-on: https://gerrit.causes.com/19438
Reviewed-by: Kim Douglas Clarence Mason <kim.mason@causes.com>
Tested-by: Kim Douglas Clarence Mason <kim.mason@causes.com>